Refactor Framework Configuration and Enhance Model Loading
- Updated FrameworkConfigurationManager to utilize default configurations for LLM frameworks, improving clarity and consistency.
- Introduced GenericLLMConfiguration for better handling of generic framework configurations.
- Enhanced ModelLoader to support new model formats, including picoLLM and ggml, and improved error handling in LLMError for better context.
- Refactored LLM services to ensure consistent error handling and memory management across all services.

+++ examples/ios/RunAnywhereAI/RunAnywhereAI/Services/LLMService/BaseLLMService.swift
@@ -157,7 +157,7 @@ class BaseLLMService: LLMService {
             modelMemory: 0,
             contextMemory: 0,
             peakMemory: 0,
-            availableMemory: ProcessInfo.processInfo.physicalMemory,
+            availableMemory: Int64(ProcessInfo.processInfo.physicalMemory),
             memoryPressure: .normal,
             cacheSize: 0
         )
@@ -195,7 +195,7 @@ class BaseLLMService: LLMService {
         HealthCheckResult(
             isHealthy: isInitialized,
             frameworkVersion: frameworkInfo.version,
-            availableMemory: ProcessInfo.processInfo.physicalMemory,
+            availableMemory: Int64(ProcessInfo.processInfo.physicalMemory),
             modelLoaded: isInitialized,
             lastError: nil,
             diagnostics: [:]
